added css to specific gameId
you can now access specific game when searching in the searchbar and pressing the game u want to know about.

It sends the id to method and applies that to a route then we take the parameter which is a id and use that to find which object in database it is and save that in a variable and can later use in both front/backend.